fix retry;

+++ main.js
@@ -14,11 +14,14 @@ const dataFile = path.join(__dirname, 'data.json');
 const cacheDir = path.join(__dirname, '_Cache');
 
 (async () => {
-    // await new Jable('https://jable.tv/videos/fsdss-234/').start();
-    let text = fs.readFileSync(dataFile, 'utf-8');
-    var data = JSON.parse(text);
-    for (let k in data) {
-        await new Jable(data[k].url).start();
+    if (fs.existsSync(dataFile)) {
+        let text = fs.readFileSync(dataFile, 'utf-8');
+        var data = JSON.parse(text);
+        for (let k in data) {
+            await new Jable(data[k].url, data[k].name).start();
+        }
+    } else {
+        await new Jable('https://jable.tv/videos/dv-1303/', 'DV-1303 為隊員們處理性慾的超可愛棒球部經紀人 小島南').start();
     }
 })().catch(err => console.log(err));
 
@@ -51,51 +54,56 @@ async function sleep(n) {
     });
 }
 
-async function download(that, download_url, save_path, retry) {
-    while (retry-- > 0) {
-        const res = await fetch(download_url, headers(that.authority, that.url));
-        if (res.status === 401) {
-            fs.unlinkSync(that.cache_file);
-        }
-        if (res.status === 200) {
-            const buff = await res.buffer();
-            fs.writeFileSync(save_path, buff, 'binary');
-            break;
-        } else {
-            console.log(`download failed, ${retry} ${download_url} => ${res.status} ${res.statusText}`);
-            if (fs.existsSync(save_path)) { fs.unlinkSync(save_path); }
-            await sleep(1);
-            await download(that, download_url, save_path, retry);
-        }
-    };
+async function download(that, download_url, save_path, trys) {
+    if (trys >= 10) {
+        console.log(`download: error, now trys is ${trys} ${download_url}`);
+        return;
+    }
+    trys = trys || 1;
+    const res = await fetch(download_url, headers(that.authority, that.url));
+    if (res.status === 401) {
+        fs.unlinkSync(that.cache_file);
+    }
+    if (res.status === 200) {
+        console.log(`download: ${trys} success, ${download_url} => ${res.status} ${res.statusText}`);
+        const buff = await res.buffer();
+        fs.writeFileSync(save_path, buff, 'binary');
+    } else {
+        console.log(`download: ${trys} failed, ${download_url} => ${res.status} ${res.statusText}`);
+        if (fs.existsSync(save_path)) { fs.unlinkSync(save_path); }
+        await sleep(1);
+        await download(that, download_url, save_path, ++trys);
+    }
 }
 
-function Jable(url) {
+function Jable(url, name) {
     this.url = url;
+    this.name = name;
+    this.id = this.name.split(' ')[0];
+    this.dir = path.join(__dirname, this.id);
+    if (!fs.existsSync(this.dir)) { fs.mkdirSync(this.dir); }
+    let txt_file = `${this.id}.txt`;
+    this.txt_path = path.join(this.dir, txt_file);
+    let mp4_name = `${this.name}.mp4`;
+    this.mp4_path = path.join(this.dir, mp4_name);
     that = this;
     this.getWeb = async () => {
-        let cache_name = url.split('://')[1].replaceAll('/', '_');
+        let cache_name = that.url.split('://')[1].replaceAll('/', '_');
         that.cache_file = path.join(cacheDir, cache_name);
         let body;
         if (fs.existsSync(that.cache_file)) {
             console.log(`useCache: ${that.cache_file}`);
             body = fs.readFileSync(that.cache_file, 'utf-8');
         } else {
-            console.log(`getWeb: ${this.url}`);
-            body = await fetch(this.url, { agent: agent, headers: headers().headers }).then(res => res.text()).catch(err => console.error('getWeb', err));
+            console.log(`getWeb: ${that.url}`);
+            body = await fetch(that.url, { agent: agent, headers: headers().headers }).then(res => res.text()).catch(err => console.error('getWeb', err));
             fs.writeFileSync(that.cache_file, body, 'utf-8');
         }
         that.m3u8_url = body.split("var hlsUrl = '")[1].split("';")[0];
-        let $ = cheerio.load(body);
-        that.name = $('h4').eq(0).text();
-        that.id = that.name.split(' ')[0];
-        that.dir = path.join(__dirname, that.id);
-        if (!fs.existsSync(that.dir)) { fs.mkdirSync(that.dir); }
-        let txt_name = `${that.id}.txt`;
-        that.txt_path = path.join(that.dir, txt_name);
-        let mp4_name = `${that.name}.mp4`;
-        that.mp4_path = path.join(that.dir, mp4_name);
-        console.log(`url=${that.url} id=${that.id} name=${that.name}`);
+        if (!that.name) {
+            let $ = cheerio.load(body);
+            that.name = $('h4').eq(0).text();
+        }
     }
     this.getM3u8 = async () => {
         that.authority = URL.parse(that.m3u8_url).host;
@@ -104,7 +112,7 @@ function Jable(url) {
         let m3u8_path = path.join(that.dir, m3u8_file);
         if (!fs.existsSync(m3u8_path)) {
             console.log(`getM3u8: ${that.m3u8_url} => ${m3u8_path}`);
-            await download(that, that.m3u8_url, m3u8_path, 3);
+            await download(that, that.m3u8_url, m3u8_path);
         }
         that.parser = new Parser();
         that.parser.push(fs.readFileSync(m3u8_path));
@@ -130,7 +138,7 @@ function Jable(url) {
         let segment = that.parser.manifest.segments[i];
 
         let key = await that.getKey(segment);
-        if (!key) { throw new Error('key error'); }
+        if (!key) { throw new Error('getKey error'); }
 
         let ts_url = that.prefix + segment.uri;
         let dl_name = path.basename(ts_url, '.ts');
@@ -140,9 +148,9 @@ function Jable(url) {
         if (!fs.existsSync(ts_path)) {
             if (!fs.existsSync(dl_path)) {
                 console.log(`getTs: [${i}/${that.total}] ${ts_url} => ${dl_path}`);
-                await download(that, ts_url, dl_path, 3);
+                await download(that, ts_url, dl_path);
             }
-            if (!fs.existsSync(dl_path)) { throw new Error('dl error'); }
+            if (!fs.existsSync(dl_path)) { throw new Error('getTs error'); }
 
             await that.decrypt(key, segment, dl_path, ts_path);
         }
@@ -183,10 +191,14 @@ function Jable(url) {
         }
     }
     this.start = async () => {
+        if (that.mp4_path && fs.existsSync(that.mp4_path)) {
+            console.log(that.mp4_path);
+            return;
+        }
+        console.log(`start: url=${this.url} id=${this.id} name=${this.name}`);
         await that.getWeb();
-        if (that.mp4_path && fs.existsSync(that.mp4_path)) { return; }
         await that.getM3u8();
-        let thread = 8; // 并行下载数量
+        let thread = 10; // 并行下载数量
         let times = (that.total % thread) > 0 ? parseInt(that.total / thread) + 1 : (that.total / thread); // 要多少次
         for (let m = 0; m < times; m++) {
             let batch = []; // 一批次
@@ -196,7 +208,7 @@ function Jable(url) {
                     batch.push(that.getTs(i));
                 }
             }
-            await Promise.all(batch); // 等所有并行完成
+            await Promise.all(batch); // 等一批次所有并行完成
         }
         await that.merge();
     }
